// apps/server/functions/src/app/graphql.schema.interface.ts
export enum AuthUserResultOperationType {
    REGISTER_AUTH_USER = "REGISTER_AUTH_USER",
    SIGNIN_AUTH_USER = "SIGNIN_AUTH_USER"
}
// ...
export interface AuthUserInput {
    email: string;
    password: string;
}

export interface CreateUserTokenInput {
    userId: string;
    accountId: string;
}

export interface NewAccountUserInput {
    email: string;
    password: string;
    accountId: string;
}
// ...
export interface AuthUserResponse {
    __typename?: 'AuthUserResponse';
    status: string;
    operationType: AuthUserResultOperationType;
    operationResult: string;
    statusCode: number;
    idToken?: Nullable<string>;
}
// ...
type Nullable<T> = T | null;